+++ containers/changelog.md
@@ -7,6 +7,9 @@
 * Add `compareSize` for `IntSet` and `IntMap`. (Soumik Sarkar)
   ([#1135](https://github.com/haskell/containers/pull/1135))
 
+* Add `mapMaybe` for `Seq`, `Set` and `IntSet`. (Phil Hazelden)
+  ([#1159](https://github.com/haskell/containers/pull/1159)
+
 ### Performance improvements
 
 * Improved performance for `Data.IntMap.restrictKeys` and